requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.4
beautifulsoup4==4.9.3
bidict==0.21.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.2
django-allauth==0.44.0
djangorestframework==3.12.4
dnspython==1.16.0
docopt==0.6.2
enum-compat==0.0.3
eventlet==0.25.2
greenlet==0.4.16
idna==2.10
Js2Py==0.71
monotonic==1.5
oauthlib==3.1.0
packaging==20.9
pipreqs==0.4.10
psycopg2-binary==2.8.6
pycparser==2.20
pyjsparser==2.7.1
PyJWT==2.1.0
pyparsing==2.4.7
PyPrind==2.11.3
pySmartDL==1.3.4
python-engineio==4.1.0
python-socketio==5.2.1
python3-openid==3.2.0
pytz==2020.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.4
yarg==0.1.9
zope.event==4.5.0
zope.interface==5.4.0
